% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SingleFunctionTaxaContributionPlots.R
\name{SingleFunctionTaxaContributionPlots}
\alias{SingleFunctionTaxaContributionPlots}
\title{Plots one of the following plots for a single function:
- Steps plot with mean and std contributions for all taxa
- Ring plot where the contributions are shown as a ring
- Bar plot where the contributions are shown as a bar}
\usage{
SingleFunctionTaxaContributionPlots(input_dir = NULL, input_prefix = NULL,
  input_suffix = ".tab", input_score = "wilcoxon",
  input_matrix_mean = "mean_stat", input_matrix_std = "std_stat",
  input_permutation = "multi_taxa", input_original = "original_value",
  input_function = NULL, min_contribution = 0, input_function_meta = NULL,
  input_taxa_taxonomy = NULL, plot_type = "steps",
  flip_case_control = FALSE, class_label_1 = NULL, class_label_2 = NULL,
  add_phyla_names_to_species_label = FALSE)
}
\arguments{
\item{input_dir}{The directory where all the FiShTaCo input files can be found (default: NULL).}

\item{input_prefix}{The prefix for all input files (e.g., pathway_with_t2f, default: NULL).}

\item{input_suffix}{The suffix for all input files (default: ".tab").}

\item{input_score}{The calculated shift score for all input files (default: wilcoxon).}

\item{input_matrix_mean}{The mean contribution stat name (default: "mean_stat").}

\item{input_matrix_std}{The contribution std stat name (default: "std_stat").}

\item{input_permutation}{The permutation type used for all input files (default: "multi_taxa").}

\item{input_original}{The original shift stat name (default: "original_value").}

\item{input_function}{The function to plot (default: NULL).}

\item{min_contribution}{The minimum taxon contribution to include in plot  (default: 0).}

\item{input_function_meta}{Metadata file for the functions to include (default: NULL).}

\item{input_taxa_taxonomy}{Taxonomy file for the given taxa (default: NULL).}

\item{plot_type}{defines the type of plot to be plotted (default: "steps").}

\item{flip_case_control}{flip the case and control contribution (default: FALSE).}

\item{class_label_1}{label for the the controls (default: NULL).}

\item{class_label_2}{label for the the cases (default: NULL).}

\item{add_phyla_names_to_species_label}{indicating whether to add the phylum to each species (default: FALSE).}
}
\value{
a handle to the resulting plot.
}
\description{
Plots one of the following plots for a single function:
- Steps plot with mean and std contributions for all taxa
- Ring plot where the contributions are shown as a ring
- Bar plot where the contributions are shown as a bar
}

